+++ .gitignore
@@ -66,3 +66,4 @@ report.[0-9]*.[0-9]*.[0-9]*.[0-9]*.json
 *.swp
 *.swo
 *.swn
+
