+++ backend/src/api/article.controller.ts
@@ -45,5 +45,10 @@ export class ArticlesController {
   @Get('status/unmoderated')
   async findUnmoderatedArticles(): Promise<Article[]> {
     return this.articlesService.findByStatus('Unmoderated');
-}
+  }
+
+  @Get('status/moderated')
+  async findArticlesArticles(): Promise<Article[]> {
+    return this.articlesService.findByStatus('Moderated');
+  }
 }
