+++ frontend/src/app/pages/analyse/[id]/page.tsx
@@ -1,0 +1,166 @@
+'use client';
+
+import React, { useState, useEffect } from "react";
+import { useParams, useRouter } from 'next/navigation';
+
+interface Article {
+  _id: string;
+  title: string;
+  authors: string;
+  source: string;
+  publicationYear: string;
+  doi: string;
+  rating: number;
+  status: string;  
+}
+
+const AnalyzePage = () => {
+
+  const router = useRouter();
+  const { id } = useParams(); // Get the article ID from the URL
+  const [displayedArticle, setDisplayedArticle] = useState<Article | null>(null); // Store the first article
+  const [loadingArticle, setLoadingArticle] = useState<boolean>(true);
+  const [loadingSubmit, setLoadingSubmit] = useState<boolean>(false);
+
+  const articleId = Array.isArray(id) ? id[0] : id;
+
+  useEffect(() => {
+		if (articleId) {
+			fetchArticle(articleId);
+		}
+	}, [articleId]);
+
+  const fetchArticle = async (id: string) => {
+    try {
+      setLoadingArticle(true);
+      const response = await fetch(`http://localhost:8082/articles/${id}`);
+      if (!response.ok) {
+        throw new Error('Network response was not ok');
+      }
+
+      const articleData: Article = await response.json();
+      if (articleData && articleData.status === 'Moderated') {
+        setDisplayedArticle(articleData);
+      } else {
+        alert('This article is not available for moderation');
+        router.push('/pages/analyze');
+      }
+    } catch (error) {
+      console.error('Failed to fetch article: ', error);
+    } finally {
+      setLoadingArticle(false);
+    }
+  };
+
+  const onSubmit = async () => {
+    setLoadingSubmit(true);
+    try {
+      const response = await fetch(`http://localhost:8082/articles/${displayedArticle!._id}`, {
+        method: 'PUT',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({
+          ...displayedArticle,
+          status: 'Analyzed',
+        }),
+      });
+
+      if (response.ok) {
+        console.log('Article updated successfully');
+        router.push('/pages/analyze');
+      } else {
+        console.error('Failed to update');
+      }
+    } catch (error) {
+      console.error('Error updating article: ', error);
+    } finally {
+      setLoadingSubmit(false);
+    }
+  };
+
+  const onReject = async () => {
+    setLoadingSubmit(true);
+    try {
+      const response = await fetch(`http://localhost:8082/articles/${displayedArticle!._id}`, {
+        method: 'PUT',
+        headers: {
+          'Content-Type': 'application/json'
+        },
+        body: JSON.stringify({
+          ...displayedArticle,
+          status: 'Rejected',
+        }),
+      });
+
+      if (response.ok) {
+        console.log('Article updated successfully');
+        router.push('/');
+      } else {
+        console.error('Failed updating article');
+      } 
+    } catch (error) {
+      console.error('Error updating article:', error);
+    } finally {
+      setLoadingSubmit(false);
+    }
+  };
+
+  if (loadingArticle) {
+    return <p>Loading...</p>; 
+  } 
+
+  if (!displayedArticle) {
+    return <p>No article found.</p>; // Show message if no article is found
+  }
+
+  return (
+    <>
+      <p>Moderation Page</p>
+
+      <br />
+
+      <div>
+        <h1>{displayedArticle.title}</h1>
+        <h3>{displayedArticle.authors}</h3>
+        <h3>{displayedArticle.source}, {displayedArticle.publicationYear}</h3>
+        <h3>
+          <a href={displayedArticle.doi} target="_blank" rel="noopener noreferrer">{displayedArticle.doi}</a>
+        </h3>
+        <h4>Article ID: {displayedArticle._id}</h4>
+      </div>
+
+      <br />
+
+      <div className="flex space-x-4">
+        <button 
+          onClick={onReject}
+          // !! Disabled button style is pretty ugly
+          className={`px-4 py-2 font-semibold rounded ${
+            loadingSubmit
+              ? 'bg-red-300 text-red-700 cursor-not-allowed'  // Disabled styles
+              : 'bg-red-500 text-white hover:bg-red-600'      // Active styles
+          }`}
+          disabled={loadingSubmit}
+        >
+          Reject
+        </button>
+        <button 
+          onClick={onSubmit}
+          className={`px-4 py-2 font-semibold rounded ${
+            loadingSubmit
+              ? 'bg-blue-300 text-blue-700 cursor-not-allowed'  // Disabled styles
+              : 'bg-blue-500 text-white hover:bg-blue-600'      // Active styles
+          }`}
+          disabled={loadingSubmit}
+        >
+          Submit for Analysis
+        </button>
+      </div>
+
+    </>
+  ) 
+
+};
+
+export default AnalyzePage;
